# PROJET_JAVA/Biblioth/src/main/java/com/example/Biblioth/Auth/AuthResponse.java
package com.example.Biblioth.Auth;

public class AuthResponse {
    private String token;
    private String role;

    public AuthResponse(String token, String role) {
        this.token = token;
        this.role = role;
    }

    public String getToken() {
        return token;
    }

    public String getRole() {
        return role;
    }
}